+++ Sources/Puppy/Compatibility.swift
@@ -1,0 +1,5 @@
+// THIS FILE IS ONLY NEEDED FOR the SDK less than macOS 13.0.
+import Foundation
+
+extension Date: @unchecked Sendable { }
+extension URL: @unchecked Sendable { }