+++ Sources/Puppy/ConsoleLogger.swift
@@ -1,8 +1,20 @@
+@preconcurrency import Dispatch
 import Foundation
 
-public class ConsoleLogger: BaseLogger {
+public final class ConsoleLogger: Loggerable, Sendable {
+    public let label: String
+    public let queue: DispatchQueue
+    public let logLevel: LogLevel
+    public let logFormat: LogFormattable?
 
-    public override func log(_ level: LogLevel, string: String) {
+    public init(_ label: String, logLevel: LogLevel = .trace, logFormat: LogFormattable? = nil) {
+        self.label = label
+        self.queue = DispatchQueue(label: label)
+        self.logLevel = logLevel
+        self.logFormat = logFormat
+    }
+
+    public func log(_ level: LogLevel, string: String) {
         print(string)
     }
 }