+++ Sources/Puppy/LogFormatter.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-public protocol LogFormattable {
+public protocol LogFormattable: Sendable {
     func formatMessage(_ level: LogLevel, message: String, tag: String, function: String, file: String, line: UInt, swiftLogInfo: [String: String], label: String, date: Date, threadID: UInt64) -> String
 }
 
@@ -10,6 +10,7 @@ extension LogFormattable {
     }
 }
 
+@Sendable
 public func dateFormatter(_ date: Date, locale: String = "en_US_POSIX", dateFormat: String = "yyyy-MM-dd'T'HH:mm:ss.SSSZZZZZ", timeZone: String = "") -> String {
     let dateFormatter = DateFormatter()
     dateFormatter.locale = Locale(identifier: locale)